availiação da tela de login

+++ src/Graxei/Graxei.Negocio.Contrato/IServicoCidades.cs
@@ -5,10 +5,10 @@ namespace Graxei.Negocio.Contrato
 {
     public interface IServicoCidades : IServicoEntidades<Cidade>
     {
-        Cidade Get(string nome, int idEstado);
+        Cidade Get(string nome, long idEstado);
         Cidade Get(string nome, Estado estado);
         IList<Cidade> Get(Estado estado);
-        IList<Cidade> GetPorEstado(int idEstado);
+        IList<Cidade> GetPorEstado(long idEstado);
         IList<Cidade> GetPorSiglaEstado(string sigla);
         IList<Cidade> GetPorNomeEstado(string nome);
     }
