+++ src/Graxei/Graxei.Persistencia.Contrato/IRepositorioBairros.cs
@@ -6,11 +6,11 @@ namespace Graxei.Persistencia.Contrato
     public interface IRepositorioBairros : IRepositorioEntidades<Bairro>
     {
         IList<Bairro> Get(Cidade cidade);
-        IList<Bairro> Get(int idCidade);
-        IList<Bairro> GetPorCidade(string nomeCidade, int idEstado);
+        IList<Bairro> Get(long idCidade);
+        IList<Bairro> GetPorCidade(string nomeCidade, long idEstado);
         Bairro Get(string nomeBairro, string nomeCidade, Estado estado);
-        Bairro Get(string nomeBairro, string nomeCidade, int idEstado);
-        Bairro Get(string nomeBairro, int idCidade);
+        Bairro Get(string nomeBairro, string nomeCidade, long idEstado);
+        Bairro Get(string nomeBairro, long idCidade);
         Bairro Get(string nomeBairro, Cidade cidade);
     }
 }
